DEVELOP - Introduce market currencies module with model, repository, and service for managing market currency data. Update markets module to integrate market currency creation during market setup. Enhance main.go to include routes for market management and update relevant data types for consistency.

+++ modules/markets/dto.go
@@ -1,15 +1,15 @@
 package markets
 
 type RunnerGet struct {
-	RunnerName        string `json:"runner_name" binding:"required"`
-	DisplayRunnerName string `json:"display_runner_name" binding:"required"`
-	Odds              uint16 `json:"odds" binding:"required"`
+	RunnerName        string  `json:"runner_name" binding:"required"`
+	DisplayRunnerName string  `json:"display_runner_name" binding:"required"`
+	Odds              float32 `json:"odds" binding:"required"`
 }
 
 type CreateMarketDTO struct {
 	MarketName string      `json:"market_name" binding:"required"`
 	Runners    []RunnerGet `json:"runners" binding:"required"`
-	Min        uint16      `json:"min" binding:"required"`
-	Max        uint16      `json:"max" binding:"required"`
-	MaxProfit  uint16      `json:"max_profit" binding:"required"`
+	Min        float32     `json:"min" binding:"required"`
+	Max        float32     `json:"max" binding:"required"`
+	MaxProfit  float32     `json:"max_profit" binding:"required"`
 }
